% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-RcppML.R
\name{method-nmf}
\alias{method-nmf}
\alias{basis,nmf-method}
\alias{coef,nmf-method}
\title{Methods for RcppML nmf}
\usage{
\S4method{basis}{nmf}(object, ...)

\S4method{coef}{nmf}(object, ...)
}
\arguments{
\item{object}{A \link[RcppML:nmf]{nmf} object.}

\item{...}{Not used currently.}
}
\description{
Add \link{NMF} package methods for RcppML \link[RcppML:nmf]{nmf} object.
}
\seealso{
\itemize{
\item \link[NMF:basis-coef-methods]{basis}
\item \link[NMF:basis-coef-methods]{coef}
}
}
