% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpnmf-methods.R
\name{mpnmf-method}
\alias{mpnmf-method}
\alias{print.mpnmf}
\alias{mp_programs}
\alias{mp_identity}
\alias{mp_scores}
\alias{mp_signatures}
\alias{mp_samples}
\alias{mp_coverage}
\title{Methods for \code{mpnmf} object.}
\usage{
\method{print}{mpnmf}(x, s_min = 1/3, ...)

mp_programs(x, s_min = 1/3)

mp_identity(x, s_min = 1/3, flatten = FALSE)

mp_scores(x, s_min = 1/3)

mp_signatures(x, s_min = 1/3, n_signatures = 20L)

mp_samples(x, s_min = 1/3)

mp_coverage(x, s_min = 1/3)
}
\arguments{
\item{x}{A \link[=mp]{mpnmf} object.}

\item{s_min}{A scalar integer indicates the minimal number of samples or a
scalar numeric (\verb{0 < s_min < 1}) indicates the minimal proportion of samples
to define the meta program. If \code{NULL}, no filters will be applied. Default:
\code{1/3}.}

\item{...}{Not used currently.}

\item{flatten}{A boolean value indicates whether to return an atomic
characters instead of a list.}

\item{n_signatures}{A scalar integer to specify the number of features to
define the program signature.}
}
\description{
\code{mp_signatures}: A list of features to define the meta program signature.
}
\details{
\code{mp_samples}: A list of samples indicates where the meta program is from.

\code{mp_coverage}: A numeric indicates the fraction of samples the meta program
was detected.

\code{mp_identity}: A list or an atomic character indicates  which program the
cells were allocated.
}
