% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpnmf-package.R
\docType{package}
\name{mpnmf-package}
\alias{mpnmf}
\alias{mpnmf-package}
\title{mpnmf: Meta Program from NMF (mpnmf)}
\description{
Identify meta programs from NMF results.
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
