% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mp_heatmap}
\alias{mp_heatmap}
\title{Meta program similarity heatmap}
\usage{
mp_heatmap(
  x,
  ...,
  col = viridisLite::viridis(100, option = "A", direction = -1),
  palette = NULL,
  show_row_names = FALSE,
  show_column_names = FALSE
)
}
\arguments{
\item{x}{A \link[=mp]{mpnmf} object.}

\item{...}{Additional arguments passed to \link[ComplexHeatmap:Heatmap]{Heatmap}.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{palette}{A named color define to meta program bar color in top and
left. Can also be a string of palette name in
\link[RColorBrewer:ColorBrewer]{RColorBrewer}.}

\item{show_row_names}{Whether show row names.}

\item{show_column_names}{Whether show column names.}
}
\description{
Meta program similarity heatmap
}
